# trefle/config/openapis.yml
openapi: 3.0.0
info:
  title: Api moteur de financement formation professionnelle
  description: Obtenir la liste des financements possibles (coûts et rémunération) en fonction du profil du bénéficiaire et des caractéristiques de la formation.
  version: 1.1.0
  contact:
    email: david.foucher@pole-emploi.fr
servers:
  - url: http://trefle.pole-emploi.fr
    description: serveur de production
  - url: http://trefle.beta.pole-emploi.fr
    description: serveur de recette
tags:
- name: financement
  description: Obtenir la liste des financements possibles (coûts et rémunération) en fonction du profil du bénéficiaire et des caractéristiques de la formation.
paths:
  /financement: &baseendpoint
    post:
      summary: Lancer un calcul de financement
      parameters:
        - name: eligible
          in: query
          required: false
          description: Permet de filtrer les financements selon leur éligibilité
          schema:
            type: boolean
        - name: explain
          in: query
          required: false
          description: Permet d'obtenir l'explication de l'application des règles dans chacun des financements
          schema:
            type: boolean
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/request'
            examples:
              compute-request:
                $ref: '#/components/examples/compute-financement-request'
      responses:
        200:
          description: Calcul effectué avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/financement-collection'
              examples:
                compute-success:
                  $ref: '#/components/examples/compute-financement-success'
        400:
          description: Le format de la requête ou l'information est invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
  /api-moteur/0.7/financement: *baseendpoint
  /api-moteur/0.8/financement: *baseendpoint
  /remuneration-aide: &baseremuendpoint
    post:
      summary: Lancer un calcul de remuneration
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/request-remuneration'
            examples:
              compute-request:
                $ref: '#/components/examples/compute-remuneration-request'
      responses:
        200:
          description: Calcul effectué avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/remuneration-collection'
              examples:
                compute-success:
                  $ref: '#/components/examples/compute-remuneration-success'
        400:
          description: Le format de la requête, ou l'information  est invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errors'
  /api-moteur/0.8/remuneration: *baseremuendpoint
components:
  schemas:
    request:
      properties:
        beneficiaire:
          $ref: '#/components/schemas/beneficiaire'
        formation:
          $ref: '#/components/schemas/formation'
    request-remuneration:
      properties:
        beneficiaire:
          $ref: '#/components/schemas/beneficiaire'
        formation:
          $ref: '#/components/schemas/formation'
        financement:
          $ref: '#/components/schemas/financement'
    financement-collection:
      required:
        - financements
      properties:
        financements:
          $ref: '#/components/schemas/financements'
    financements:
      type: array
      items:
        $ref: "#/components/schemas/financement"
    remuneration-collection:
      required:
        - remunerations
      properties:
        remunerations:
          $ref: '#/components/schemas/remunerations'
    remunerations:
      type: array
      items:
        $ref: "#/components/schemas/remuneration"
    errors:
      type: object
      additionalProperties:
        type: string
  examples:
    compute-financement-request:
      summary: Données du bénéfiaire et formation
      value:
        {
          "beneficiaire": {
           "droit_prive": false,
           "inscrit_pe": true,
           "naissance": "2000-10-12"
           },
          "formation": {
            "numero": "03_1702030F"
          }
        }
    compute-financement-success:
      summary: Liste des financements éligible
      value:
        {
          "financements": [
            {
              "type_lbf": "cpf",
              "priorite_lbf": 0,
              "tags": [
                 "CPF",
                 "DE",
                 "individuel",
                 "national"
              ],
              "cpf_mobilisable": true,
              "financee_pe": false,
              "cumulable": true,
              "intitule": "Compte Personnel de Formation (CPF)",
              "eligible": true,
              "explain": null,
              "reste_a_charge": 0,
              "prise_en_charge": null,
              "plafond_prix_horaire": 0,
              "plafond_prise_en_charge": 0,
              "remuneration": 0,
              "fin_remuneration": null,
              "indemnite_conges_payes": 0,
              "heures": 0,
              "prise_en_charge_texte": "Formation totalement ou partiellement financée dans la limite de votre solde cpf et à hauteur de 15 euros de l'heure. Dans le cas d'un financement partiel, vous pouvez prendre en charge vous-même le reste du montant ou faire appel à des cofinancements. Un abondement de votre compte est possible dans certains cas, demandez les conditions à votre conseiller en évolution professionnelle.",
              "demarches": "1. Contactez votre conseiller référent emploi pour valider avec lui votre projet de formation.\n2. Consultez votre compte CPF sur le site https://www.moncompteactivite.gouv.fr",
              "description": "Le CPF peut financer totalement ou partiellement une formation certifiante. NB : dans le cas d'une formation cofinancée par la région ou Pôle emploi, le compte personnel de formation (CPF) pourra être mobilisé.",
              "en_savoir_plus": "",
              "remuneration_texte": ""
            }
          ]
        }
    compute-remuneration-request:
      summary: Données du bénéfiaire et formation
      value:
        {
          "beneficiaire": {
                  "age": 40,
                  "naissance": "1980-01-01",
                  "allocation": 300,
                  "type_allocation": "ass",
                  "th": true,
                  "a_travaille_douze_mois_sur_24_mois": true,
                  "a_travaille_douze_mois": true,
                  "a_travaille_six_mois": true,
                  "parent_isole": false,
                  "mere_trois_enfant": false,
                  "veuve_ou_divorcee": false,
                  "seule_et_enceinte": false,
                  },
            "formation": {
                           "codes_financeur": [2],
                           "region": 27,
                        },
            "financement": {
              "identifiant": "dispositif 6"
            }
          }
    compute-remuneration-success:
      summary: Rémunération et aides éligibles
      value:
        {
          "remunerations": [
            {
              "remuneration": 910.02,
              "fin_remuneration": null,
              "indemnite_conges_payes": 0,
              "intitule_remuneration": "Aides et rémunération mobilisable par la région Bourgogne Franche Comté",
              "droit_aide_complementaire": true,
              "montant_aide_complementaire": 200.0,
              "droit_aide_transport": true,
              "description_aide_transport": "L'indemnité de transport est limitée à un aller/retour sur la durée de la formation.",
              "droit_aide_hebergement": true,
              "montant_aide_hebergement": "101,84€ par mois (sous condition)",
              "description_aide_hebergement": "Si la distance vers le lieu de votre formation est supérieure à 250km. Alors vous pouvez bénéficier d'une aide à l'hébergement. Cette aide est non cumulable avec l'aide au transport."
            }
          ]
        }